Extract abstract superclass AbstractPerson for PersonCat & PersonDog;
